Extracted facebook event data parsing to a dedicated class
